Added select and lenses helpers.
These, along with the `pr` QuasiQuoter offer nicer inline syntax for
projecting specific columns out of a row.

Changed the types of the generated primed lenses so that they preserve
the column name. This might come in handy in the future if we want to
use these lens values to plumb types around, but they are quite hard to
work with due to the amount of polymorphism (the functors and full row
type).