More expressive data parsing for type inference.
Using the new Parseable class, we can indicate whether a value could not
be parsed, if it could be parsed with likely ambiguity, or it could be
parsed without ambiguity. This is particularly useful for types that
have a Maybe aspect to them that may be serialized as an empty
string. Since many types can be represented with an empty string, we
want the non-empty strings to drive the type inference.