Be more willing to assign the Text type to a column
If we don’t do so, and discard rows containing data that does not fit
into a narrower column type, we lose too much.

This does mean that we are now pickier about assigning a column a
numeric type where previously we might have discarded rows with non
numeric data in what might be a numerically-typed column.

For instance, the weather data set used by the Kata04 demo program
includes some temperature readings with an asterisk
suffix (e.g. "93*"). Previously, rows with such data were discarded
since the column was assigned a type of ‘Double‘ (most rows have
integers in the columns in question, but there are some with decimal
points). Now we assign such a column a type of ‘Text‘, so the user
must handle the non-numeric rows either on the Haskell side after
parsing, or as a pre-processing step.

Since Frames produces a bunch of useful definitions based on the CSV
data itself, some pre-processing is probably the right move.

An example of needing to be more willing to type a column as ‘Text‘ is
the ‘Mpg.hs‘ test program and associated ‘mpg.csv‘ data file’s ‘drv‘
column. This indicates the drive of a car, and might be "4", "f", or
"r". One could imagine an ’Int’, and a ’Bool’ or an ’Int’ for the
first two options, but the third requires that the column be ’Text’.

Fixes #86