Add table joins via Data.Discrimination (#105)
* Add inner join to start

* Fix imports

* Add Joins module to export

* Add joins demo to cabal

With profiling enabled

* Instances cleanup for joins

* joins code cleanup

* Fix performance problem with join

All I needed to swap was foldl for foldr. Now
about 30-40 times slower than dplyr on an equivalent
problem. On paper this should scale better, but closing
this gap for smallish problems would be nice.

* Create a dropCols function (anti-select)

* Haddock inner join

* Funs to make Rec Maybe's

Necessary for outerjoins

* Add outer join

A bit painful that it returns a list of records now in
the Maybe interpretation functor. I think ultimately I'd
like them to still be Records with the internal types wrapped
in Maybe...

* Add left and right outer joins

* Fix haddock

* Code cleanup, thanks hlint

* Add simple check for innerJoin

Checks that the join is impervious to
reordering the right table. Hand testing shows
left, right, and outers performing as expected.
Performance problems also appear to be overstated.
Now within an easy factor of 10 of dplyr.

Next steps checking string joins and multicolumn
joins.

* Add an grouping instance for text

* Elaborate the joins demo

* Make joins demo a proper demo

* Use TypeApplications for joins

Requires AllowAmbiguousTypes as well

* Add explicit exports to Joins

* Switch to using TypeApplications for joins

Includes updated demo

* Fix mergeRightEmpty

These joins weren't actually working, the key was getting dropped
when merging a right with a Nothing left.

* Improve haddock substantially

* Upcase j in joinCols in join examples

Cosmetic, but all the column types will have first letter capitalized names.

* Fix typo

Remove quotes in `JoinColN` in code examples

* Add pandas joins comparison