vinyl-0.8 compatibility
This shows an annoying regression in vinyl usability due to the recent
extra parameterization of the RSubset type class. It would be good if
we could re-order the type parameters on the class to make picking the
right instance work the way it used to (where you could @subset to
pick out the fields you want).