Don't require an odd number of quotation marks
When we check if a CSV line part is a quoted string, we look for an
odd number of quotation marks at both the beginning and end. This does
not support nested quotes. This change fixes #114, but unintentional
breakage elsewhere is possible as this is clearly a fragile bit of parsing logic.