Eliminate clash with vinyl's rget and rput
This renames `Frames.Rec.rget` to `rgetField` and `Frames.Rec.rput` to
`rputField`. This makes it easier when you have a  non-`Identity`
functor and wish to use vinyl's `rget` function.

Move the Uncurry* demo programs into the test directory and include
them as part of the spec test suite.

Translate the README from markdown to org so that we can directly
include the Uncurry* example programs. As these programs are now also
run by the test suite, this should make it less likely that the README
silently falls out of date. We now generate the markdown file from the
org file with the `ox-gfm` package's `org-gfm-export-to-markdown` function.