Generate simple sum types for categorical variables
The default is to generate a data type with up to 8 constructors for a
column, but this is relatively easy to override in user
code (i.e. `Categorical 16` as a part of your column universe for up
to 16 variants).

In order to not break existing code, you need to specifically
opt in to a column universe with categorical types. The test cases
demonstrate one way of doing so.