Express record parsing in terms of tokenized input
The idea is to push Frames functionality further downstream so that it
takes a stream of tokenized rows, rather than a stream of rows, as input.

This further decouples us from the separated value parsing part of
things. We still include a CSV parser for the sake of simplicity and a
smaller dependency footprint, but we also now work with external data
file parsers such as that provided by the new `Frames-dsv`
package (built atop `hw-dsv`).