Remove nameless attributes from adult.csv test data
The data file downloaded from the host website includes an extra
column on each row encoding an attribute for an example labeling
task. This extra column is not part of the nominal data; there is no
column heading for it. We do not use it in our demo programs, so we
strip this column when downloading the file.