Use a larger column type inference prefix in the benchdemo program
It was pointed out by Eric Lindblad that the benchdemo output was
similar but not identical to the comparison pandas output. This was
due to an incorrect type inference of a column of the data file used
as the first 1000 lines of the file are insufficient to observe the
variety of values that column should be able to represent.